fix rounding

+++ app/src/main/java/com/breadwallet/wallet/wallets/bitcoin/WalletBitcoinManager.java
@@ -396,7 +396,7 @@ public class WalletBitcoinManager extends BRCoreWalletManager implements BaseWal
     @Override
     public BigDecimal getFiatBalance(Context app) {
         BigDecimal bal = getFiatForSmallestCrypto(app, new BigDecimal(getCachedBalance(app)), null);
-        return new BigDecimal(bal == null ? 0 : bal.longValue());
+        return new BigDecimal(bal == null ? 0 : bal.doubleValue());
     }
 
     @Override
